% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add}
\usage{
add(a, b)
}
\arguments{
\item{a}{number value}

\item{b}{number value}
}
\value{
sum of a and b
}
\description{
Example function that adds two values together.
}
\examples{
add(5, 6)
# 11
}
