% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependant.R
\name{dependant}
\alias{dependant}
\title{Dependant}
\usage{
dependant()
}
\value{
None
}
\description{
Plots a line using ggplot2.
Within a package library(...) or require(...) are not used.
}
\examples{
dependant()
# Plots line in viewer.
}
